refactor(vended-tools): treat bash as its own deprecated tool
Follow-up feedback on the rename: it should read as a new shell tool plus a
deprecated bash tool, one tool per home, rather than one tool with a
shape-shifting name.

Python moves the deprecated surface (bash instance, make_bash, rationale) out
of shell/shell.py into vended_tools/_bash.py, leaving the shell module
canonical-only. The module is _bash rather than bash because importing a
public bash submodule would bind it onto the package, silently shadowing the
__getattr__ that emits the DeprecationWarning and handing callers a module
where they expect a tool.

TypeScript gains ShellOutput in shell/types.ts, aliasing the bash tool's
output shape and mirroring the Python SDK's ShellOutput, so the shell module
no longer reaches into ../bash for its return type.
