+++ strands-py/src/strands/vended_tools/__init__.py
@@ -23,10 +23,10 @@ Example Usage:
 import warnings
 from typing import Any
 
+from ._bash import _RENAME_RATIONALE, make_bash  # noqa: F401  deprecated tool, kept importable until v2.0.0
 from .file_editor import file_editor, make_file_editor
 from .http_request import http_request, make_http_request
 from .shell import make_shell, shell
-from .shell.shell import _RENAME_RATIONALE, make_bash  # noqa: F401  deprecated alias, kept importable until v2.0.0
 from .sleep import make_sleep, sleep
 
 
@@ -39,7 +39,7 @@ def __getattr__(name: str) -> Any:
             DeprecationWarning,
             stacklevel=2,
         )
-        from .shell.shell import bash
+        from ._bash import bash
 
         return bash
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
