+++ strands-ts/src/vended-tools/shell/index.ts
@@ -12,3 +12,4 @@
 export { makeShell, makeBash } from './make-shell.js'
 export type { MakeShellOptions, MakeBashOptions } from './make-shell.js'
 export { SANDBOX_SHELL_DESCRIPTION, ShellTimeoutError, ShellExecutionError } from './types.js'
+export type { ShellOutput } from './types.js'
