+++ strands-ts/src/vended-tools/shell/types.ts
@@ -5,7 +5,7 @@
  * callers who caught the pre-rename error types keep working.
  */
 
-import { BashTimeoutError, BashSessionError } from '../bash/types.js'
+import { BashTimeoutError, BashSessionError, type BashOutput } from '../bash/types.js'
 
 export const SANDBOX_SHELL_DESCRIPTION =
   'Executes shell commands. Each call runs in a fresh shell; ' +
@@ -36,3 +36,10 @@ export class ShellExecutionError extends BashSessionError {
     this.name = 'ShellExecutionError'
   }
 }
+
+/**
+ * Output format for shell command execution. The same shape as the bash tool's
+ * output, so pre-rename consumers keep working; mirrors the Python SDK's
+ * `ShellOutput`.
+ */
+export type ShellOutput = BashOutput
